backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
loguru>=0.7.2
tenacity>=8.2.3

# Authentication and security
PyJWT>=2.8.0
cryptography>=41.0.0
fastapi-clerk-auth>=0.0.7

# Database dependencies   
sqlalchemy>=2.0.25

# CopilotKit and Research
copilotkit
exa-py==1.9.1
httpx>=0.27.2,<0.28.0

# AI/ML dependencies
openai>=1.3.0
anthropic>=0.7.0
mistralai>=0.0.12
google-genai>=1.0.0
google-ai-generativelanguage>=0.6.18,<0.7.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0

# Web scraping and content processing
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
html5lib>=1.1
aiohttp>=3.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# SEO Analysis dependencies
advertools>=0.14.0
textstat>=0.7.3
pyspellchecker>=0.7.2
aiofiles>=23.2.0
crawl4ai>=0.2.0

# Image and audio processing for Stability AI
Pillow>=10.0.0
scikit-learn>=1.3.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pydantic>=2.5.2,<3.0.0
typing-extensions>=4.8.0

# Optional dependencies (for enhanced features)
redis>=5.0.0
schedule>=1.2.0 